requirements.txt
numpy
scipy
tqdm
scikit-learn
tensorflow
jax
jaxlib
optax
tensorcircuit-nightly